% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genevintage-package.R
\docType{package}
\name{genevintage-package}
\alias{genevintage}
\alias{genevintage-package}
\title{genevintage: Detect Annotation Releases and Repair Gene Names}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Detects the species and the 'Ensembl' \url{https://www.ensembl.org} or 'GENCODE' \url{https://www.gencodegenes.org} annotation release associated with a set of gene or transcript identifiers by scoring them against a bundled fingerprint of each release. Identifiers are then converted to gene names using the matching release. Gene symbols that spreadsheet software turned into dates or numbers ('SEPT9' to '9-Sep') are restored to the spelling of the chosen release. Also retrieves genes by class (mitochondrial, sex-linked, protein-coding, ribosomal, transfer RNA, long non-coding RNA), exon-union gene lengths, 'Entrez' cross-references, and orthologs in another species from 'Ensembl Compara'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genevintage.saketlab.org/R/}
  \item \url{https://github.com/saketlab/genevintage}
  \item Report bugs at \url{https://github.com/saketlab/genevintage/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

Authors:
\itemize{
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})
}

}
\keyword{internal}
