% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genevintage-package.R
\docType{package}
\name{genevintage-package}
\alias{genevintage}
\alias{genevintage-package}
\title{genevintage: Date the Annotation Vintage of a Set of Gene Identifiers}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Given a vector of gene identifiers, detects the species from the identifier prefixes and ranks the 'Ensembl' or 'GENCODE' releases consistent with the set, then converts the identifiers to gene names. Identifiers with no gene name are returned unchanged, so results stay usable as row labels. Further functions return genes by class (mitochondrial, sex-linked, protein-coding, ribosomal, transfer RNA, long non-coding RNA) and map genes to their orthologs in another species. Release fingerprints ship with the package; mapping tables are streamed from the 'Ensembl' GTF on first use and cached for offline reuse, as described at \url{https://www.ensembl.org/info/data/ftp/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saketlab/genevintage}
  \item Report bugs at \url{https://github.com/saketlab/genevintage/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

Authors:
\itemize{
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})
}

}
\keyword{internal}
