% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genevintage-package.R
\docType{package}
\name{genevintage-package}
\alias{genevintage}
\alias{genevintage-package}
\title{genevintage: Identify the Build Annotation Associated with a Set of Gene Identifiers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Genevintage detects the species from the identifier prefixes and ranks the 'Ensembl' or 'GENCODE' releases consistent with the set, then converts the identifiers to gene names. It also supports methods to retrieve gene by class (mitochondrial, sex-linked, protein-coding, ribosomal, transfer RNA, long non-coding RNA) and mapping genes to their orthologs in another species.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genevintage.saketlab.org/R/}
  \item \url{https://github.com/saketlab/genevintage}
  \item Report bugs at \url{https://github.com/saketlab/genevintage/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

Authors:
\itemize{
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})
}

}
\keyword{internal}
